# openapi.yaml
openapi: 3.0.3
info:
  title: LifeOps API
  version: "1.0"
servers:
  - url: http://localhost:8080
paths:
...
  /v1/worker/run:
    post:
      summary: Trigger worker for agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent:
                  type: string
                  description: Agent name
              required:
                - agent
      responses:
        '200':
          description: Worker executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  agent:
                    type: string
...